pyproject.toml
[tool.poetry]
name = "gi-corp"
version = "0.1.0"
description = ""
authors = ["YADHU KRISHNA P V <116164765+yadhu12345@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
